+++ test/clojure/nrepl/core_test.clj
@@ -512,3 +512,12 @@
                             first
                             :value)]
     (is (= cloned-sess-*1 "5"))))
+
+(def-repl-test print-namespace-maps-binding
+  (when (resolve '*print-namespace-maps*)
+    (let [set-true (repl-eval session "(set! *print-namespace-maps* true)")
+          true-val (first (repl-values session "*print-namespace-maps*"))
+          set-false (repl-eval session "(set! *print-namespace-maps* false)")
+          false-val (first (repl-values session "*print-namespace-maps*"))]
+      (is (= true true-val))
+      (is (= false false-val)))))